Refactor main.ts Phase 2: Extract app initialization and terminal creation (#392)
Reduces main.ts from 1174 to 997 lines (177 lines removed).

## Phase 2.1: Extract app initialization (~136 lines)
- Created src/lib/app-initializer.ts with:
  - checkDependenciesOnStartup()
  - initializeApp() - priority-based workspace loading (CLI > localStorage > Tauri storage)
- Implements dependency injection pattern for state, validators, and render
- Removed inline implementation from main.ts

## Phase 2.2: Extract terminal creation (~56 lines)
- Moved createPlainTerminal() to src/lib/terminal-actions.ts
- Updated all call sites to pass dependencies:
  - state, workspacePath, generateNextConfigId, saveCurrentConfig
- Consistent with existing terminal action patterns

## Files changed:
- **src/lib/app-initializer.ts** (NEW): 183 lines
- **src/lib/terminal-actions.ts**: +66 lines
- **src/main.ts**: -199 lines (221 removed, 22 for dependency passing)

## Testing:
- ✅ TypeScript compilation passes
- ✅ Vite build succeeds
- ✅ Biome linting passes
- ✅ Rust formatting passes
- ✅ Clippy passes
- ⚠️ One pre-existing flaky daemon test failure (unrelated to changes)
